requirements.txt
annotated-types==0.7.0
boto3==1.38.10
botocore==1.38.10
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coverage==7.8.0
cryptography==44.0.3
filelock==3.18.0
flake8==7.2.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
hf-xet==1.1.0
httplib2==0.22.0
huggingface-hub==0.31.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
moto==5.1.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.25.7
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
setuptools==80.4.0
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
xmltodict==0.14.2
